requirements.txt
aiohttp>=3.9.0
asyncio>=3.4.3
cryptography>=41.0.0
PyYAML>=6.0.1
pynput>=1.7.6
psutil>=5.9.0
PyGObject>=3.42.0
pycairo>=1.23.0
Flask>=3.0.0
Flask-CORS>=4.0.0
Werkzeug>=3.0.0
openai>=1.3.0
google-generativeai>=0.3.0
anthropic>=0.7.0
python-daemon>=3.0.1
dbus-python>=1.3.2
python-dotenv>=1.0.0
requests>=2.31.0
jsonschema>=4.20.0